requirements.txt
# OpenAI Whisper Debug Project Requirements
openai-whisper>=20231117
numpy>=1.26.0
soundfile>=0.12.0
psutil>=5.9.0
torch>=2.0.0
torchaudio>=2.0.0
